pgkernelgateway/src/test/javascript/spec/app/entities/account-type/account-type-update.component.spec.ts
/* tslint:disable max-line-length */
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder } from '@angular/forms';
import { Observable, of } from 'rxjs';

import { PgkernelgatewayTestModule } from '../../../test.module';
import { AccountTypeUpdateComponent } from 'app/entities/account-type/account-type-update.component';
import { AccountTypeService } from 'app/entities/account-type/account-type.service';
import { AccountType } from 'app/shared/model/account-type.model';

describe('Component Tests', () => {
  describe('AccountType Management Update Component', () => {
    let comp: AccountTypeUpdateComponent;
    let fixture: ComponentFixture<AccountTypeUpdateComponent>;
    let service: AccountTypeService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [PgkernelgatewayTestModule],
        declarations: [AccountTypeUpdateComponent],
        providers: [FormBuilder]
      })
        .overrideTemplate(AccountTypeUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(AccountTypeUpdateComponent);
      comp = fixture.componentInstance;
      service = fixture.debugElement.injector.get(AccountTypeService);
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', fakeAsync(() => {
        // GIVEN
        const entity = new AccountType(123);
        spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.update).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));

      it('Should call create service on save for new entity', fakeAsync(() => {
        // GIVEN
        const entity = new AccountType();
        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.create).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));
    });
  });
});
